Simplify installation
By moving the library to the root of the repository it becomes possible to use the popular "Add .ZIP Library" installation process, rather than the more complex manual installation previously required.

This change also opens the possibility of further simplifying installation and updates by adding this library to the Arduino Library Manager index.

As a bonus, this removes the terrible advice to install the library to the Arduino IDE installation folder, which will cause the library to be lost whenever updating to a different IDE version.